% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_other.R
\name{convert}
\alias{convert}
\alias{sphere_to_euclid}
\alias{euclid_to_sphere}
\title{Convert Spherical/Euclidean Coordinates}
\usage{
sphere_to_euclid(x)

euclid_to_sphere(x)
}
\arguments{
\item{x}{either spherical or Euclidean input}
}
\description{
Convert Spherical/Euclidean Coordinates
}
