% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_sm.R
\name{g_disk}
\alias{g_disk}
\title{Euclidean distance on disk}
\usage{
g_disk(z, dV, r = NULL, axis = 3)
}
\arguments{
\item{z}{matrix of euclidean coordinates}

\item{dV}{boundary in euclidean coordinates}

\item{r}{radius of sphere}

\item{axis}{axis on which to project to, i.e. this axis is kept constant and distance is measured from the remaining two}
}
\value{
List containing two elements
\item{\code{g}}{values of function \code{g}}
\item{\code{grad}}{matrix of derivatives of \code{g}}
}
\description{
Boundary function which projects region onto 2D disk
}
\details{
For data in Euclidean coordinates on a sphere, such as those from von-Mises Fisher distribution,
this function projects these points onto a 2D plane, and takes the Euclidean distance between the now 2D
points and the boundary.

To specify \code{axis}, i.e. the axis of projection, include \code{axis} as an element of
the \code{options} list in the call to \code{\link{sphere_sm}},
e.g. \code{sphere_sm(..., options = list(axis = 2))}.

This function is used for describing the boundary function \eqn{g}.
}
