% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{mvn_dist}
\alias{mvn_dist}
\alias{psi_mvn}
\alias{mvn}
\title{Derivatives of the Log Density of Multivariate Normal Distribution}
\usage{
psi_mvn(theta, x, Sigma = NULL)

mvn(Sigma = NULL)
}
\arguments{
\item{theta}{parameters to be estimated, can include elements of the covariance matrix}

\item{x}{random variable x, taken one row at a time}

\item{Sigma}{optional; either a fully supplied covariance matrix, or a single digit indicating the diagonal of the covariance matrix. If not supplied, the covariance matrix is estimated.}
}
\value{
list containing two elements
\item{\code{f}}{the evaluation of the first derivative}
\item{\code{grad}}{the evaluation of the second derivative}
}
\description{
The first and second derivative of the log density of the Multivariate Normal Distribution
}
