% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{polygon_points}
\alias{polygon_points}
\title{Make Polygon Points}
\usage{
polygon_points(x, y, res = 20)
}
\arguments{
\item{x}{vector of \code{x} vertices that form a closed cycle}

\item{y}{vector of \code{y} vertices that form a closed cycle}

\item{res}{resolution to set between vertices, i.e. number of points in each line segment}
}
\value{
a matrix whose first column are the points in \code{x} and whose second column are the points in \code{y}
}
\description{
Create a series of points in a polygon with a given resolution
}
\details{
\code{x} and \code{y} are vectors whose pairs give the vertices of the polygon object

This is a simple function that takes the same initial two arguments as \code{\link[graphics]{polygon}}, but creates
points that go over the course of the polygon edges. To create a full shape, the final vertex of the pairs \code{x}
and \code{y} must match the first vertex.
}
\examples{
# make unit square
  polygon_points(c(-1,-1,1,1,-1),c(-1,1,1,-1,-1))
}
