% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{project}
\alias{project}
\title{Project data to Lq ball}
\usage{
project(x, q = 2, mod = 1)
}
\arguments{
\item{x}{data matrix n x d}

\item{q}{order of Lq ball}

\item{mod}{size of ball}
}
\value{
projections as coordinates
}
\description{
Project data to Lq ball
}
