% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{psi_mvn}
\alias{psi_mvn}
\title{Derivatives of the Log Density of Multivariate Normal Distribution}
\usage{
psi_mvn(theta, x)
}
\arguments{
\item{theta}{parameters to be estimated, can include elements of the covariance matrix}

\item{x}{random variable x, taken one row at a time}
}
\value{
list containing two elements
\item{\code{f}}{the evaluation of the first derivative}
\item{\code{grad}}{the evaluation of the second derivative}
}
\description{
The first and second derivative of the log density of the Multivariate Normal Distribution
}
