% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_sm.R
\name{psi_vmf}
\alias{psi_vmf}
\title{Derivatives of the Log Density of von-Mises Fisher distribution}
\usage{
psi_vmf(mu, x, k = 10)
}
\arguments{
\item{mu}{mean direction to be estimated}

\item{x}{random variable x, taken one row at a time}

\item{k}{concentration parameter, default \code{k=10}}
}
\value{
list containing two elements
\item{\code{f}}{the evaluation of the first derivative}
\item{\code{grad}}{the evaluation of the second derivative}
}
\description{
The first and second derivative of the log density of the von-Mises Fisher distribution
}
