% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_other.R
\name{radian}
\alias{radian}
\title{Convert Degrees to radian}
\usage{
radian(x)
}
\arguments{
\item{x}{input in degrees}
}
\value{
\code{x} but in radians instead
}
\description{
Take array of Colatitude, Longitude in degrees and convert to radians
}
