% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_sm.R
\name{sphere_sm}
\alias{sphere_sm}
\title{Truncated Spherical Score Matching}
\usage{
sphere_sm(
  x,
  dV,
  family = "vmf",
  g,
  init = colMeans(euclid_to_sphere(x)),
  options = list()
)
}
\arguments{
\item{x}{spherical coordinate data}

\item{dV}{boundary coordinates}

\item{family}{distribution, e.g. "vmf" for von-Mises Fisher}

\item{g}{boundary function}

\item{init}{optional; initial condition}

\item{options}{optional list; non-specified \code{family}
by including \code{psi} which returns first and second derivative of log pdf as \code{f} and \code{grad}}
}
\value{
parameter estimates
}
\description{
Truncated Spherical Score Matching
}
\details{
if \code{g} is not included as an argument, then non-truncated score matching will be used. For
a sphere, use
\describe{
\item{\code{"Haversine"}}{haversine distance function \code{\link{g_hav}}}
\item{\code{"Euclidean"}}{euclidean projection distance function \code{\link{g_disk}}}
}
}
