% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_sm.R
\name{spherical_dist}
\alias{spherical_dist}
\alias{psi_vmf}
\alias{psi_kent}
\title{Derivatives of the Log Density of spherical distributions}
\usage{
psi_vmf(mu, x, k = 10)

psi_kent(par, x, k = 10, beta = 1)
}
\arguments{
\item{mu}{mean direction to be estimated for the von-Mises Fisher distribution}

\item{x}{random variable x, taken one row at a time}

\item{k}{concentration parameter, default \code{k=10}}

\item{par}{matrix of parameters to estimate for the Kent distribution}

\item{beta}{ovalness parameter, default \code{beta=1}}
}
\value{
list containing two elements
\item{\code{f}}{the evaluation of the first derivative}
\item{\code{grad}}{the evaluation of the second derivative}
}
\description{
The first and second derivative of the log density of the von-Mises Fisher distribution and Kent distribution
}
\details{
\code{psi_vmf} corresponds to the von-Mises Fisher distribution, where \code{psi_kent} corresponds to the
Kent distribution. The parameter to be estimated for the von-Mises Fisher distribution is the mean direction \code{mu},
whereas for the Kent distribution, the parameters are \code{mu}, and the major and minor axes. The parameters enter
\code{psi_kent} in a matrix with columns corresponding to these parameters respectively.
}
