% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{truncsm}
\alias{truncsm}
\title{Truncated Score Matching}
\usage{
truncsm(x, dV, family = mvn(), init = NULL, options = list(), ...)
}
\arguments{
\item{x}{truncated random observation}

\item{dV}{boundary where \code{x} was truncated (in points)}

\item{family}{distribution family to look in, currently supports multivariate normal as "mvn"}

\item{init}{initial conditions to be passed to \code{\link[stats]{optim}}}

\item{options}{extra options that can be specified, see details}

\item{...}{additional arguments that can be passed to optim}
}
\value{
parameter estimates given by score matching
}
\description{
For truncated data within some boundary, estimate parameters that do not necessarily rely on the boundary.
}
\details{
For details of the procedure, see "Estimating Density Models with Complex Truncation Boundaries" by Song Liu et al.

The variable \code{x} takes values within a certain truncated region, where the boundary of this region is defined by
\code{dV}. It is possible through score matching to estimate parameters that are not constrained
to the boundary by minimising the difference in score functions of the model and the data - the gradient of the log pdf, i.e.
\deqn{
\psi = \nabla log p(x; \theta).
}

If \code{family} is not supplied, then \code{psi} needs to be supplied as an element of \code{options}.
\code{psi} will be function will calculates the derivatives of the log pdf, and needs to output a list
containing two elements: the first derivative as \code{f} and second derivative as \code{grad}.

All arguments to \code{options} are:
\describe{
\item{\code{psi}}{function outputting first and second derivatives of log pdf as \code{f} and \code{grad}, taking two inputs only: \code{theta} and \code{x}; the parameter to estimate over and the truncated observations.}
\item{\code{g}}{function outputting function controlling behaviour at the boundary, outputting value and first derivative of the function as \code{g} and \code{grad}, taking two inputs only: the truncated observations, \code{x} and the boundary points \code{dV}.}
}
See \code{\link{psi_mvn}} and \code{\link{g_def}} for examples.
}
