% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{g_def}
\alias{g_def}
\title{Default Boundary Function}
\usage{
g_def(x, dV)
}
\arguments{
\item{x}{single point in truncated region}

\item{dV}{boundary in points}
}
\value{
list with two elements
\item{\code{g}}{values of \code{g}}
\item{\code{grad}}{matrix of derivatives of \code{g}}
}
\description{
Default boundary function \eqn{g_0(x)} as described in "Estimating Density Models with Complex Truncation Boundaries"
by Song Liu et al
}
