% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_sm.R
\name{g_disk}
\alias{g_disk}
\title{Euclidean distance on disk}
\usage{
g_disk(z, dV)
}
\arguments{
\item{z}{matrix of euclidean coordinates}

\item{dV}{boundary in euclidean coordinates}
}
\value{
List containing two elements
\item{\code{g}}{values of function \code{g}}
\item{\code{grad}}{matrix of derivatives of \code{g}}
}
\description{
Boundary function which projects region onto 2D disk
}
\details{
For data in Euclidean coordinates on a sphere, such as those from von-Mises Fisher distribution,
this function projects these points onto a 2D disk, and takes the Euclidean distance between the now 2D
points and the boundary.

This is used for describing the boundary function \eqn{g}.
}
