% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_sm.R
\name{g_hav}
\alias{g_hav}
\title{Haversine distance on sphere}
\usage{
g_hav(z, dV, r = 1)
}
\arguments{
\item{z}{matrix of euclidean coordinates}

\item{dV}{boundary in euclidean coordinates}

\item{r}{radius of sphere, default 1}
}
\value{
List containing two elements
\item{\code{g}}{values of function \code{g}}
\item{\code{grad}}{matrix of derivatives of \code{g}}
}
\description{
Calculate great circle distance between a set of points and a boundary, with gradient
}
