% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_sm.R
\name{spherical_dist}
\alias{spherical_dist}
\alias{psi_vmf}
\alias{psi_kent}
\alias{kent}
\alias{vmf}
\title{Derivatives of the Log Density of spherical distributions}
\usage{
psi_vmf(par, x, k = NULL)

psi_kent(par, x, k = NULL, b = NULL)

kent(k = NULL, b = NULL)

vmf(k = NULL)
}
\arguments{
\item{par}{parameter vector to be estimated}

\item{x}{random variable x, taken one row at a time}

\item{k}{concentration parameter (vmf and kent distribution)}

\item{b}{ovalness parameter (kent distribution)}
}
\value{
list containing two elements
\item{\code{f}}{the evaluation of the first derivative}
\item{\code{grad}}{the evaluation of the second derivative}
}
\description{
The first and second derivative of the log density of the von-Mises Fisher distribution and Kent distribution
}
\details{
\code{psi_vmf} corresponds to the von-Mises Fisher distribution, where \code{psi_kent} corresponds to the
Kent distribution. The parameters to be estimated for the von-Mises Fisher distribution are the mean direction \code{mu} and
concentration parameter \code{k}. Whereas for the Kent distribution, the parameters are \code{mu}, the major and minor axes
\code{gamma1} and \code{gamma2}, the concentration parameter \code{k} and ovalness parameter \code{b}. The parameters enter
\code{psi_kent} in a long vector.
}
